Add files via upload
In an effort to drastically reduce the Cost of Engineering (COE) in each new project, Volvo implemented the theory of CAST; Common Architecture, Similar Technology.
In this publication, my team was tasked to create and document the use of a "Common Cab." Wheel Loaders, Motor Graders, Soil Compactors and Asphalt Compactors all offered Cab enclosures for safety and operator comfort, yet utilized three (3) different SKUs for the cabs, and four (4) unique interfaces for operations. In this example, we were provided a Cab for a Wheel Loader, and instructed to "Make it work" on a Soil Compactor, regardless of size.
This project was much more fun than it sounds, and really helped put our team on the map as an under-utilized resource. The changes were adopted company-wide.