Add files via upload
Although my title with Volvo was "Technical Writer II",  my role was much more than that.  When need arose, I would work as the actual "Methods Technician" as well; the tech that got dirty.  It was a natural fit, as I had transferred from Operations (Assembly) to Engineering.  For this Motor Grader service manual, I was the Tech, Writer, Illustrator, and Publisher.  I include "Publisher" to note that role is also responsible for Translation Integrity; hence, a German sample.  Shortly after the publication of this manual, Volvo exited the Motor Grader market. The section covers the hydraulic functions, specifications, installation and repair of the HTE1160 transmission/hydraulic brake accumulator.